+++ src/App.tsx
@@ -1,7 +1,7 @@
 import './App.css'
 import Header from "./components/Header/Header";
 import Footer from "./components/Footer/Footer";
-import {SelectArea, SelectOption} from "./components/DropDown/SelectBox";
+import QNABoard from "./pages/Q&ABoard/Q&ABoard";
 function App(){
   return (
       <div className="App">
@@ -9,22 +9,7 @@ function App(){
               <Header/>
           </header>
           <main className={'bg-[#F5F5F5]'}>
-              <div className={'p-12'}>
-                  <SelectArea>
-                      <SelectOption>
-                          aa
-                      </SelectOption>
-                      <SelectOption>
-                          bb
-                      </SelectOption>
-                      <SelectOption>
-                          cc
-                      </SelectOption>
-                      <SelectOption>
-                          dd
-                      </SelectOption>
-                  </SelectArea>
-              </div>
+              <QNABoard/>
           </main>
           <footer className={'flex justify-between items-center p-2 w-full border-t border-t-[#D9D9D9]'}>
               <Footer/>
