+++ src/pages/Q&ABoard/Q&ABoard.tsx
@@ -3,7 +3,8 @@ import SortIcon from "../../components/Icon/Sort";
 import SearchIcon from "../../components/Icon/Search";
 import CategoryIcon from "../../components/Icon/category";
 import TableRow from "../../components/TableRow/TableRow";
-import DropDownRe from "../../components/DropDown/Remake";
+import {DropDown} from "../../components/DropDown/DropDown";
+import {SelectOption} from "../../components/DropDown/SelectBox";
 
 export default function QNABoard(){
 
@@ -13,7 +14,7 @@ export default function QNABoard(){
                 Q&A 게시판
             </h1>
             <div className={'w-full border-y border-y-[#EAEAEA] bg-white h-fit'}>
-                <div className={'sticky bg-primary px-2 py-2 flex items-center justify-between gap-2'}>
+                <div className={'bg-primary px-2 py-2 flex items-center justify-between gap-2'}>
                     <button className={'flex justify-center items-center bg-primary focus:outline-none px-2 py-1 border-2 border-white rounded-lg w-[50px]'}>
                         <PenIcon/>
                     </button>
@@ -21,30 +22,34 @@ export default function QNABoard(){
                         <button className={'flex justify-center items-center bg-primary focus:outline-none px-2 py-1 border-2 border-white rounded-lg w-[50px]'}>
                             <SearchIcon/>
                         </button>
-                        <button className={'flex justify-center items-center bg-primary focus:outline-none px-2 py-1 border-2 border-white rounded-lg w-[50px]'}>
-                            <CategoryIcon/>
-                        </button>
+                        <DropDown component={
+                                <span className={'flex justify-center items-center bg-primary focus:outline-none px-2 py-1 border-2 border-white rounded-lg w-[50px]'}>
+                                    <CategoryIcon/>
+                                </span>
+                                }
+                                  location={'right'}
+                        >
+                            <SelectOption value={'카테고리1'}>
+                                adawdawadwadwadw
+                            </SelectOption>
+                            <SelectOption value={'카테고리2'}>
+                                a
+                            </SelectOption>
+                            <SelectOption value={'카테고리3'}>
+                                a
+                            </SelectOption>
+                        </DropDown>
                         <button className={'flex justify-center items-center bg-primary focus:outline-none px-2 py-1 border-2 border-white rounded-lg w-[50px]'}>
                             <SortIcon/>
                         </button>
-                        <DropDownRe>
-                            <option value={'1'}>옵션1</option>
-                            <option value={'1'}>옵션1</option>
-                            <option value={'1'}>옵션1</option>
-                            <option value={'1'}>옵션1</option>
-                            <option value={'1'}>옵션1</option>
-                            <option value={'1'}>옵션1</option>
-                        </DropDownRe>
                     </div>
                 </div>
                 <div className={'bg-white min-h-[500px] p-3 flex flex-col gap-3'}>
-                    <TableRow img={'/logo512.png'} title={'제목'} writer={'작성자'} date={'2020-12-31'} view={2020} comment={10}/>
-                    <TableRow img={'/logo512.png'} title={'제목'} writer={'작성자'} date={'2020-12-31'} view={2020} comment={10}/>
-                    <TableRow img={'/logo512.png'} title={'제목'} writer={'작성자'} date={'2020-12-31'} view={2020} comment={10}/>
-                    <TableRow img={'/logo512.png'} title={'제목'} writer={'작성자'} date={'2020-12-31'} view={2020} comment={10}/>
-                    <TableRow img={'/logo512.png'} title={'제목'} writer={'작성자'} date={'2020-12-31'} view={2020} comment={10}/>
-                    <TableRow img={'/logo512.png'} title={'제목'} writer={'작성자'} date={'2020-12-31'} view={2020} comment={10}/>
-                    <TableRow img={'/logo512.png'} title={'제목'} writer={'작성자'} date={'2020-12-31'} view={2020} comment={10}/>
+                    {
+                        Array(20).fill(
+                            <TableRow img={'/logo512.png'} title={'제목'} writer={'작성자'} date={'2020-12-31'} view={2020} comment={10}/>
+                        )
+                    }
                 </div>
                 <div className={'flex justify-between items-center p-3'}>
                     <button className={'flex justify-center items-center bg-primary focus:outline-none px-2 py-2 border-2 border-white rounded-lg w-[50px]'}>
